+++ core/runtime/src/main/java/mb/pie/runtime/store/InMemoryStoreBase.java
@@ -210,6 +210,7 @@ public abstract class InMemoryStoreBase implements Store, StoreReadTxn, StoreWri
             }
             setTaskObservability(key, data.taskObservability);
         } else {
+            this.taskInputs.remove(key);
             clearInternalObject(key);
             this.taskOutputs.remove(key);
             this.taskObservability.remove(key);
@@ -220,6 +221,12 @@ public abstract class InMemoryStoreBase implements Store, StoreReadTxn, StoreWri
                 addTaskRequire(key, dep.callee);
                 addTaskRequireDep(key, dep);
             }
+        } else {
+            // Ensure that `key` is removed from `callersOf`, even though `callersOf.get(key)` may return an empty list,
+            // because an empty list has different semantics than not being in the mapping at all: if
+            // `callersOf.get(key)` returns an empty list, then key is eligible for garbage collection through
+            // `getTasksWithoutCallers`, but that would fail because this task does not exist anymore.
+            this.callersOf.remove(key);
         }
         removeResourceRequireDepsOf(key);
         if(data != null) {
@@ -247,7 +254,7 @@ public abstract class InMemoryStoreBase implements Store, StoreReadTxn, StoreWri
         }
         final @Nullable Observability observability = this.taskObservability.remove(key);
         // Pass `true` to `removeTaskRequireDepsOf` to remove `key` from the `callersOf` map, as `key` will be removed
-        // completely, thus it is not possible to call it any more.
+        // completely, thus it is not possible to call it anymore.
         final @Nullable Collection<TaskRequireDep> removedTaskRequires = removeTaskRequiresAndDepsOf(key);
         removeFromCallersOf(key);
         final @Nullable Collection<ResourceRequireDep> removedResourceRequires = removeResourceRequireDepsOf(key);
@@ -268,16 +275,21 @@ public abstract class InMemoryStoreBase implements Store, StoreReadTxn, StoreWri
 
 
     private @Nullable Collection<TaskRequireDep> removeTaskRequiresAndDepsOf(TaskKey key) {
-        this.taskRequires.remove(key);
-        final @Nullable Collection<TaskRequireDep> removedDeps = this.taskRequireDeps.remove(key);
-        if(removedDeps != null) {
-            for(final TaskRequireDep removedDep : removedDeps) {
-                final @Nullable Set<TaskKey> callers = this.callersOf.get(removedDep.callee);
+        final @Nullable Collection<TaskKey> removedCallers = this.taskRequires.remove(key);
+        // Use the removed tasks from `taskRequires` instead of the removed deps from `taskRequireDeps` here, because
+        // when a task requires (calls) another task, this is immediately recorded in `taskRequires`, but only after
+        // the call completes in `taskRequireDeps`. This matters because tasks can be cancelled/interrupted, and in that
+        // case no dependency will be added to `taskRequireDeps`. Therefore, we should use `taskRequires` to remove
+        // entries from `callersOf`.
+        if(removedCallers != null) {
+            for(final TaskKey removedCaller : removedCallers) {
+                final @Nullable Set<TaskKey> callers = this.callersOf.get(removedCaller);
                 if(callers != null) {
                     callers.remove(key);
                 }
             }
         }
+        final @Nullable Collection<TaskRequireDep> removedDeps = this.taskRequireDeps.remove(key);
         return removedDeps;
     }